feat: swap main => new main page
