+++ src/components/EndpointEditor/EndpointEditor.tsx
@@ -21,7 +21,7 @@ const EndpointEditor = () => {
         </div>
       ) : (
         <input
-          className="pl-2 my-5 w-full min-w-[400px] h-[40px] bg-white rounded text-black outline-none"
+          className="pl-2 my-5 w-full  h-[40px] bg-white rounded text-black outline-none"
           value={inputValue}
           onChange={(e) => dispatch(setUrl(e.target.value))}
         />
